+++ R/DSAggregate_Sample.R
@@ -16,32 +16,29 @@
 # with this program; if not, write to the Free Software Foundation, Inc.,
 # 51 Franklin Street, Fifth Floor, Boston, MA 02110-1301 USA.
 
-
-
-
 #' Sampling from a Data Stream (Data Stream Operator)
 #'
 #' Extracts a sample form a data stream using Reservoir Sampling.
 #'
-#' If \code{biased=FALSE} then the reservoir sampling algorithm by McLeod and
+#' If `biased = FALSE` then the reservoir sampling algorithm by McLeod and
 #' Bellhouse (1983) is used. This sampling makes sure that each data point has
 #' the same chance to be sampled. All sampled points will have a weight of 1.
 #' Note that this might not be ideal for an evolving stream since very old data
 #' points have the same chance to be in the sample as newer points.
 #'
-#' If \code{bias=TRUE} then sampling prefers newer points using the modified
+#' If `bias = TRUE` then sampling prefers newer points using the modified
 #' reservoir sampling algorithm 2.1 by Aggarwal (2006). New points are always
 #' added. They replace a random point in thre reservoir with a probability of
-#' reservoir size over \code{k}. This an exponential bias function of
-#' \eqn{2^{-lambda}} with \eqn{lambda=1/k}.
+#' reservoir size over `k`. This an exponential bias function of
+#' \eqn{2^{-lambda}} with \eqn{lambda = 1 / k}.
 #'
-#' @family DSO
+#' @family DSAggregate
 #'
 #' @param k the number of points to be sampled from the stream.
-#' @param biased if \code{FALSE} then a regular (unbiased) reservoir sampling
+#' @param biased if `FALSE` then a regular (unbiased) reservoir sampling
 #' is used. If true then the sample is biased towards keeping more recent data
 #' points (see Details section).
-#' @return An object of class \code{DSO_Sample} (subclass of \code{DSO}).
+#' @return An object of class `DSAggregate_Sample` (subclass of `DSAggregate`).
 #' @author Michael Hahsler
 #' @references Vitter, J. S. (1985): Random sampling with a reservoir. ACM
 #' Transactions on Mathematical Software, 11(1), 37-57.
@@ -53,49 +50,63 @@
 #' Evolution. International Conference on Very Large Databases (VLDB'06).
 #' 607-618.
 #' @examples
+#' set.seed(1500)
 #'
-#' stream <- DSD_Gaussians(k=3, noise=0.05)
+#' stream <- DSD_Gaussians(k = 3, noise = 0.05)
 #'
-#' sample <- DSO_Sample(k=20)
+#' sample <- DSAggregate_Sample(k = 50)
 #'
 #' update(sample, stream, 500)
 #' sample
 #'
-#' # plot points in sample
+#' # apply kmeans clustering to the sample
+#' km <- kmeans(get_points(sample), centers = 3)
 #' plot(get_points(sample))
-#'
+#' points(km$centers, col = "red", pch = 3, cex = 2)
 #' @export
-DSO_Sample <- function(k = 100, biased = FALSE)
-  structure(list(description =
-      if(biased) "Reservoir sampling (biased)" else "Reservoir sampling",
-    RObj = SampleDSO$new(k = k, biased = biased)),
-    class = c("DSO_Sample","DSO"))
+DSAggregate_Sample <- function(k = 100, biased = FALSE)
+  structure(
+    list(
+      description =
+        if (biased)
+          "Reservoir sampling (biased)"
+      else
+        "Reservoir sampling",
+      RObj = SampleDSAggregate$new(k = k, biased = biased)
+    ),
+    class = c("DSAggregate_Sample", "DSAggregate", "DST")
+  )
 
 #' @export
-update.DSO_Sample <- function(object, dsd, n=1, verbose=FALSE, ...) {
+update.DSAggregate_Sample <-
+  function(object,
+    dsd,
+    n = 1,
+    verbose = FALSE,
+    ...) {
+    ### some matrix to be processed in one go
+    if (!is(dsd, "DSD")) {
+      n <- nrow(dsd)
+      dsd <- DSD_Memory(dsd)
+    }
 
-  ### some matrix to be processed in one go
-  if(!is(dsd, "DSD")) {
-    n <- nrow(dsd)
-    dsd <- DSD_Memory(dsd)
+    ### FIXME: we do not need to get all points if n is very large!
+    object$RObj$update(get_points(dsd, n = n), verbose = verbose, ...)
   }
 
-  ### FIXME: we do not need to get all points if n is very large!
-  object$RObj$update(get_points(dsd, n=n), verbose=verbose, ...)
-}
-
 #' @export
-get_points.DSO_Sample <- function(x, ...) {
+get_points.DSAggregate_Sample <- function(x, ...) {
   x$RObj$get_points(...)
 }
 
 #' @export
-get_weights.DSO_Sample <- function(x, ...) {
+get_weights.DSAggregate_Sample <- function(x, ...) {
   x$RObj$get_weights(...)
 }
 
 
-SampleDSO <- setRefClass("SampleDSO",
+SampleDSAggregate <- setRefClass(
+  "SampleDSAggregate",
   fields = list(
     k	= "integer",
     biased = "logical",
@@ -104,11 +115,8 @@ SampleDSO <- setRefClass("SampleDSO",
   ),
 
   methods = list(
-    initialize = function(
-      k	= 100L,
-      biased = FALSE
-    ) {
-
+    initialize = function(k	= 100L,
+      biased = FALSE) {
       k	<<- as.integer(k)
       biased	<<- biased
       stream_size	<<- 0L
@@ -123,32 +131,36 @@ SampleDSO <- setRefClass("SampleDSO",
     ### biased: more recent values have a higher probability
 
     update = function(x, ...) {
-      if(!is(data, class(x)) && !is.null(data))
+      if (!is(data, class(x)) && !is.null(data))
         stop("Data stream data type not compatible!")
 
-      if(is.data.frame(x)) update_data.frame(x)
-      else update_list(x)
+      if (is.data.frame(x))
+        update_data.frame(x)
+      else
+        update_list(x)
     },
 
     update_data.frame = function(x, ...) {
-
-      if(!biased) {
+      if (!biased) {
         ### fast initialization
-        if(is.null(data)) {
-          if(nrow(x) <= k) data <<- x
-          else data <<- x[sample(1:nrow(x), k),]
+        if (is.null(data)) {
+          if (nrow(x) <= k)
+            data <<- x
+          else
+            data <<- x[sample(1:nrow(x), k), ]
           stream_size <<- nrow(x)
-        }else{
-
+        } else{
           ### reservoir sampling
-          for(i in 1:nrow(x)){
+          for (i in 1:nrow(x)) {
             ### fill with values first
-            if(nrow(data) < k) {
-              data <<- rbind(data, x[i,])
+            if (nrow(data) < k) {
+              data <<- rbind(data, x[i, ])
 
-            }else{ ### replace values with decreasing probabilities
-              r <- sample.int(stream_size+1L, size=1L)
-              if(r <= k) data[r, ] <<- x[i,]
+            } else{
+              ### replace values with decreasing probabilities
+              r <- sample.int(stream_size + 1L, size = 1L)
+              if (r <= k)
+                data[r,] <<- x[i, ]
               ### Note: we do not need to replace weight (is already 1)
             }
 
@@ -156,16 +168,18 @@ SampleDSO <- setRefClass("SampleDSO",
           }
         }
 
-      }else{ ### biased
-        if(is.null(data)) data <<- data.frame()
+      } else{
+        ### biased
+        if (is.null(data))
+          data <<- data.frame()
 
-        for(i in 1:nrow(x)){
+        for (i in 1:nrow(x)) {
           ### add all new points and replace point in reservoir with prob=size/k
-          prob <- nrow(data)/k
-          if(sample(c(TRUE, FALSE), 1L, prob=c(prob, 1-prob))) {
-            data[sample.int(nrow(data), 1L),] <<- x[i,]
-          }else{
-            data <<- rbind(data, x[i,])
+          prob <- nrow(data) / k
+          if (sample(c(TRUE, FALSE), 1L, prob = c(prob, 1 - prob))) {
+            data[sample.int(nrow(data), 1L), ] <<- x[i, ]
+          } else{
+            data <<- rbind(data, x[i, ])
           }
 
           stream_size <<- stream_size + 1L
@@ -174,24 +188,26 @@ SampleDSO <- setRefClass("SampleDSO",
     },
 
     update_list = function(x, ...) {
-
-      if(!biased) {
+      if (!biased) {
         ### fast initialization
-        if(is.null(data)) {
-          if(length(x) <= k) data <<- x
-          else data <<- x[sample(1:nrow(x), k)]
+        if (is.null(data)) {
+          if (length(x) <= k)
+            data <<- x
+          else
+            data <<- x[sample(1:nrow(x), k)]
           stream_size <<- length(x)
-        }else{
-
+        } else{
           ### reservoir sampling
-          for(i in 1:length(x)){
+          for (i in 1:length(x)) {
             ### fill with values first
-            if(length(data) < k) {
+            if (length(data) < k) {
               data <<- append(data, x[i])
 
-            }else{ ### replace values with decreasing probabilities
-              r <- sample.int(stream_size+1L, size=1L)
-              if(r <= k) data[r] <<- x[i]
+            } else{
+              ### replace values with decreasing probabilities
+              r <- sample.int(stream_size + 1L, size = 1L)
+              if (r <= k)
+                data[r] <<- x[i]
               ### Note: we do not need to replace weight (is already 1)
             }
 
@@ -199,15 +215,17 @@ SampleDSO <- setRefClass("SampleDSO",
           }
         }
 
-      }else{ ### biased
-        if(is.null(data)) data <<- list()
+      } else{
+        ### biased
+        if (is.null(data))
+          data <<- list()
 
-        for(i in 1:length(x)){
+        for (i in 1:length(x)) {
           ### add all new points and replace point in reservoir with prob=size/k
-          prob <- length(data)/k
-          if(sample(c(TRUE, FALSE), 1L, prob=c(prob, 1-prob))) {
+          prob <- length(data) / k
+          if (sample(c(TRUE, FALSE), 1L, prob = c(prob, 1 - prob))) {
             data[sample.int(nrow(data), 1L)] <<- x[i]
-          }else{
+          } else{
             data <<- append(data, x[i])
           }
 
@@ -217,26 +235,37 @@ SampleDSO <- setRefClass("SampleDSO",
     },
 
     get_points = function(...) {
-      if(!is.null(data)) data else data.frame()
-      },
+      if (!is.null(data))
+        data
+      else
+        data.frame()
+    },
 
     get_weights = function(...) {
-      n <- if(is.null(data)) 0L else
-        if(is.data.frame(data)) nrow(data) else length(data)
+      n <- if (is.null(data))
+        0L
+      else
+        if (is.data.frame(data))
+          nrow(data)
+      else
+        length(data)
       rep(1, n)
     }
   )
 )
 
 
-### DSC interface to SampleDSO
-SampleDSC <- setRefClass("SampleDSC",
-  contains="SampleDSO",
+### DSC interface to SampleDSAggregate
+SampleDSC <- setRefClass(
+  "SampleDSC",
+  contains = "SampleDSAggregate",
 
   methods = list(
-    cluster = function(x, ...) update(x, ...),
-    get_microclusters = function(...) get_points(...),
-    get_microweights = function(...) get_weights(...)
+    cluster = function(x, ...)
+      update(x, ...),
+    get_microclusters = function(...)
+      get_points(...),
+    get_microweights = function(...)
+      get_weights(...)
   )
 )
-